Did it the right way
Created repositories and services.
