+++ app/Controllers/NewController.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace App\Controllers;
+
+
+class NewController
+{
+
+}
