+++ app/Views/CurrencyView.php
@@ -1,0 +1,5 @@
+<h1>Currencies</h1>
+
+<p>Current currency rates</p>
+
+<?php echo implode("<br>", $currencies) ?>
