+++ app/Views/IndexView.php
@@ -1,0 +1,5 @@
+<h1>Currencies</h1>
+
+<a href="./currency">
+    <p>Get current currencies</p>
+</a>
